Rename track_prepare to track_context in ground truthing.
The module builds per-track annotation context (spline, spectrogram,
audible-axis bounds), not generic track preparation.
